:recycle: use query params instead of routes
